Split FunctionalSpec and TensorSpec
